refactor: home pages 정리
