+++ src/features/notification/model/useGetNotifications.ts
@@ -1,6 +1,8 @@
-import { isLoggedIn } from '@/features/auth/model/authSlice';
 import { API_END_POINT, QUERY_KEYS, httpClient } from '@/shared';
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query';
+
+import { isLoggedIn } from '@/features/auth/model/authSlice';
+import { EmptyError } from '@/shared/lib';
 import { useSelector } from 'react-redux';
 import type { INotification } from '../config';
 
@@ -9,9 +11,10 @@ export const useGetNotifications = () => {
 
   const getNotifications = async (): Promise<INotification[]> => {
     const response = await httpClient.get(`${API_END_POINT.NOTIFICATIONS}`);
-    if (!response.data || !response.data.items) {
-      throw new Error('No items found in the response');
+    if (response.data.items.length === 0) {
+      throw new EmptyError('notification');
     }
+
     return response.data.items;
   };
 
@@ -27,9 +30,10 @@ export const useGetNotifications = () => {
 export const useGetNotificationsWithSuspense = () => {
   const getNotifications = async (): Promise<INotification[]> => {
     const response = await httpClient.get(`${API_END_POINT.NOTIFICATIONS}`);
-    if (!response.data || !response.data.items) {
-      throw new Error('No items found in the response');
+    if (response.data.items.length === 0) {
+      throw new EmptyError('notification');
     }
+
     return response.data.items;
   };
