+++ src/pages/home/lib/index.ts
@@ -1,0 +1,1 @@
+export { useScrollDetection } from './useScrollDetection';
