+++ src/pages/home/model/useGetBestAuctions.ts
@@ -1,0 +1,9 @@
+import { QUERY_KEYS } from '@/shared';
+import { useSuspenseQuery } from '@tanstack/react-query';
+import { getBestAuctions } from '../api';
+
+export const useGetBestAuctions = () => {
+  const { data: bestAuctions } = useSuspenseQuery({ queryKey: [QUERY_KEYS.BEST_AUCTIONS], queryFn: getBestAuctions });
+
+  return { bestAuctions };
+};
