+++ src/pages/home/model/useGetPreAuctions.ts
@@ -1,0 +1,12 @@
+import { QUERY_KEYS } from '@/shared/constants/queryKeys';
+import { useSuspenseQuery } from '@tanstack/react-query';
+import { getPreAuctions } from '../api';
+
+export const useGetPreAuctions = () => {
+  const { data: preAuctions } = useSuspenseQuery({
+    queryKey: [QUERY_KEYS.PRE_AUCTIONS],
+    queryFn: getPreAuctions,
+  });
+
+  return { preAuctions };
+};
