+++ src/pages/home/ui/Home.tsx
@@ -1,11 +1,10 @@
-import BestItemList from '@/components/home/BestItemList';
-import CategoryList from '@/components/home/CategoryList';
-import HomeItemField from '@/components/home/HomeItemField';
-import HomeRegisterBtn from '@/components/home/HomeRegisterBtn';
-import ImminentItemList from '@/components/home/ImminentItemList';
-import PreAuctionItemList from '@/components/home/PreAuctionItemList';
-import { useScrollDetection } from '@/hooks/useScrollDetection';
-import { LocalAPIAsyncBoundary } from '@/shared';
+import { useScrollDetection } from '../lib';
+import BestItemList from './BestItemList';
+import CategoryList from './CategoryList';
+import HomeItemField from './HomeItemField';
+import { HomeRegisterBtn } from './HomeRegisterBtn';
+import ImminentItemList from './ImminentItemList';
+import PreAuctionItemList from './PreAuctionItemList';
 
 export const Home = () => {
   const { isScrolled, elementRef } = useScrollDetection(0);
@@ -14,20 +13,14 @@ export const Home = () => {
     <div ref={elementRef} role='main' aria-label='main_area' className='relative flex flex-col justify-between w-full h-full gap-6 overflow-y-scroll'>
       <div className='flex flex-col gap-10'>
         <HomeItemField name='베스트 경매'>
-          <LocalAPIAsyncBoundary height={250}>
-            <BestItemList />
-          </LocalAPIAsyncBoundary>
+          <BestItemList />
         </HomeItemField>
         <HomeItemField name='종료 임박 경매'>
-          <LocalAPIAsyncBoundary height={250}>
-            <ImminentItemList />
-          </LocalAPIAsyncBoundary>
+          <ImminentItemList />
         </HomeItemField>
         <CategoryList />
         <HomeItemField name='사전 경매'>
-          <LocalAPIAsyncBoundary height={250}>
-            <PreAuctionItemList />
-          </LocalAPIAsyncBoundary>
+          <PreAuctionItemList />
         </HomeItemField>
       </div>
       <HomeRegisterBtn isScrolled={isScrolled} />
