+++ src/pages/home/ui/HomeItemField.tsx
@@ -1,3 +1,5 @@
+import { LocalAPIAsyncBoundary } from "@/shared";
+
 import { ReactNode } from "react";
 
 const HomeItemField = ({ children, name }: { children: ReactNode, name: string }) => {
@@ -5,7 +7,9 @@ const HomeItemField = ({ children, name }: { children: ReactNode, name: string }
     <section className='flex flex-col w-full gap-4'>
       <label className='text-heading3'>{name}</label>
       <div className='flex gap-4'>
-        {children}
+        <LocalAPIAsyncBoundary height={250}>
+          {children}
+        </LocalAPIAsyncBoundary>
       </div>
     </section>
   );
